completed all features

+++ lionnetwork/user.py
@@ -1,6 +1,7 @@
 from flask import (Blueprint, render_template, redirect, request, session, url_for, flash, g, make_response)
 from sqlalchemy.sql import text
 import uuid
+from .utils import processQuery
 from lionnetwork.auth import login_required
 
 user = Blueprint('user', __name__)
@@ -8,19 +9,53 @@ user = Blueprint('user', __name__)
 @user.route('/admin', methods = ["GET", "POST"])
 @login_required
 def adminSettings():
-    if request.method == "POST":
+    print("Called")
+    if request.method == "POST" and g.user.is_admin:
+        params = {}
         params['uni'] = request.form['uni']
-        params['admin'] = True
-        sqlUpgrade = 'update users set is_admin = :admin where columbia_uni = :uni'
-        try:
-            g.conn.execute(text(sqlUpgrade), params)
-            flash(f"Successfully upgraded {params['uni']} to admin.", category = "success")
-            return redirect(request.url)
-        except Exception as e:
-            print(e)
-            flash(f"Could not upgrade {params['uni']} to admin...", category = "error")
-            return redirect(request.url)
+        params['new_ind'] = request.form['ind']
 
+        if len(params['uni']) > 0:
+            params['admin'] = True
+            sql = 'update users set is_admin = :admin where columbia_uni = :uni'
+            flash(f"Successfully upgraded {params['uni']} to admin.", category = "success") if processQuery(sql, params) else flash(f"Could not upgrade {params['uni']} to admin...", category = "error")
+
+        elif len(params['new_ind']) > 0:
+            params['uni'] = session['columbia_uni']
+            params['new_id'] = int(str(uuid.uuid4().int)[:5])
+
+            sql = 'insert into industries (industry_id, industry_name, columbia_uni) values (:new_id, :new_ind, :uni)'
+            flash(f"Successfully added {params['new_ind']}.", category = "success") if processQuery(sql, params) else flash(f"Could not add {params['new_ind']}...", category = "error")
+
+        return redirect(request.url)
+    else:
+        return render_template('user/settings.html')
+
+@user.route('/modifyJob', methods = ["GET", "POST"])
+def modifyJob():
+    if request.method == "POST" and g.user.is_admin:
+        params = {}
+        if request.form.get('hide') is not None:
+            params['listing_id'] = int(request.form.get('hide'))
+            params['val'] = False
+            sql = 'update job_listings set show_listing = :val where listing_id = :listing_id'
+        elif request.form.get('delete') is not None:
+            params['listing_id'] = int(request.form.get('delete'))
+            sql = 'delete from job_listings where listing_id = :listing_id'
+        elif request.form.get('show') is not None:
+            params['listing_id'] = int(request.form.get('show'))
+            params['val'] = True
+            sql = 'update job_listings set show_listing = :val where listing_id = :listing_id'
+
+        flash("Query Success!", category = "success") if processQuery(sql, params) else flash("Query Failure!", category = "error")
+        return redirect(url_for('user.jobPosting'))
+
+    elif not g.user.is_admin:
+        flash("Invalid Access", category = "error")
+        return redirect(url_for('user.jobPosting'))
+
+    else:
+        return redirect(url_for('user.jobPosting'))
 
 def fetch_industries():
     industries = g.conn.execute('SELECT * FROM industries').fetchall()
@@ -33,8 +68,8 @@ def jobPosting():
         jobList = []; params = {}
         params['int'] = False; params['visa'] = False
         session['int'] = False; session['visa'] = False
-
-        params['ind'] = int(request.form['industry_id'])
+        params['ind'] = int(request.form['industry'])
+        
         if "internship" in request.form.keys():
             session['int'] = True
             params['int'] = True
@@ -66,10 +101,16 @@ def settings():
     if request.method == "GET":
         return render_template('user/settings.html')
 
-    elif request.method == "POST":
+    elif request.method == "POST" and g.user:
         params = {}
-        params['uni'] = session['columbia_uni']
-        if request.form.get('delete') is not None:
+        params['uni'] = g.user.columbia_uni
+        params['delete'] = True if "delete" in request.form.keys() else False
+        params['deactivate'] = True if "deactivate" in request.form.keys() else False
+        params['reactivate'] = True if "reactivate" in request.form.keys() else False
+        params['newName'] = request.form['name'] if "name" in request.form.keys() else None
+        params['newMajor'] = request.form['major'] if "major" in request.form.keys() else None
+
+        if params['delete']:
             sqlRemove = 'delete from users where columbia_uni = :uni'
             try:
                 g.conn.execute(text(sqlRemove), params)
@@ -78,10 +119,8 @@ def settings():
             except Exception as e:
                 print(e)
                 flash("Could not remove your account...", category = "error")
-                return redirect(request.url)
 
-        if request.form.get('deactivate') is not None:
-            params['deactivate'] = True
+        elif params['deactivate']:
             sqlDeactivate = 'update users set is_deactivated = :deactivate where columbia_uni = :uni'
             try:
                 g.conn.execute(text(sqlDeactivate), params)
@@ -90,11 +129,20 @@ def settings():
             except Exception as e:
                 print(e)
                 flash("Could not remove your account...", category = "error")
-                return redirect(request.url)
 
-        params['newName'] = request.form['name']
-        sqlName = 'update users set name = :newName where columbia_uni = :uni'
-        if len(params['newName']) > 0:
+        elif params['reactivate']:
+            print(params['deactivate'])
+            sqlReactivate = 'update users set is_deactivated = :deactivate where columbia_uni = :uni'
+            try:
+                g.conn.execute(text(sqlReactivate), params)
+                flash("Successfully reactivated your account.", category = "success")
+                return redirect(url_for('user.jobPosting'))
+            except Exception as e:
+                print(e)
+                flash("Could not reactivated your account...", category = "error")
+
+        elif len(params['newName']) > 0:
+            sqlName = 'update users set name = :newName where columbia_uni = :uni'
             try:
                 g.conn.execute(text(sqlName), params)
                 flash(f"Your name has been updated to {params['newName']}", category = "sucess")
@@ -102,9 +150,8 @@ def settings():
                 print(e)
                 flash("Could not update your name...", category = "error")
 
-        params['newMajor'] = request.form['major']
-        sqlMajor = 'update users set major = :newMajor where columbia_uni = :uni'
-        if len(params['newMajor']) > 0:
+        elif len(params['newMajor']) > 0:
+            sqlMajor = 'update users set major = :newMajor where columbia_uni = :uni'
             try:
                 g.conn.execute(text(sqlMajorName), params)
                 flash(f"Your majore has been updated to {params['newMajor']}", category = "sucess")
